requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.3.2
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
python-magic==0.4.15
six==1.12.0
tqdm==4.31.1
traitlets==4.3.2
wcwidth==0.1.7
